refactor: bookmarks의 components 폴더 아래 경로 수정 및 파일명 변경
