--- src/app/(auth)/(with-header)/bookmarks/components/bookmarkCard.type.ts
@@ -1,1 +1,0 @@
-export interface IBookmarksProps {}
