+++ src/app/(auth)/(with-header)/bookmarks/components/bookmarkCard/bookmarkCard.type.ts
